match_server refactored
