+++ match_server/src/lib/supabase/useSupabaseBasicClient.ts
@@ -1,6 +1,6 @@
 import { createClient } from "@supabase/supabase-js";
 import { Database } from "../database.types";
-import "dotenv/config"
+const dotenv = require('dotenv').config()
 
 export default function useSupabaseBasicClient() {
 
@@ -8,5 +8,5 @@ export default function useSupabaseBasicClient() {
         return createClient<Database>(process.env.SUPABASE_URL, process.env.SUPABASE_ANON_KEY);
     }
 
-    throw new Error("SUPABASE_URL or SUPABASE_ANON_KEY environment variables uninitialized.");
+    throw new Error("SUPABASE_URL or SUPABASE_ANON_KEY environment variables are undefined");
 }
